requirements.txt
catboost
scikit-learn
fastapi
uvicorn
python-multipart
pandas
numpy
joblib
google-genai